Set both VERSION and SOVERSION for target snappy
Set proper values for properties `VERSION` and `SOVERSION` of target `snappy`, so that CMake produces the link `libsnappy.so.1` to the shared library `libsnappy.so.1.1.5` that is built. This link was available in 1.1.4 (with autotools) and it is still necessary for the run time linker to work. Its absence leads to run time errors like "error while loading shared libraries: libsnappy.so.1: cannot open shared object file: No such file or directory".